optimizing includes: fix  compiling after optimizing glm includes
