+++ libs/openFrameworks/gl/ofGLRenderer.cpp
@@ -12,6 +12,7 @@
 #include "ofCamera.h"
 #include "ofTrueTypeFont.h"
 #include "ofNode.h"
+#include "ofVideoBaseTypes.h"
 
 using namespace std;
