+++ libs/openFrameworks/sound/ofSoundBuffer.cpp
@@ -9,6 +9,7 @@
 #include "ofSoundUtils.h"
 #include "ofLog.h"
 #include <limits>
+#include "glm/trigonometric.hpp"
 
 using namespace std;
 
@@ -599,7 +600,7 @@ void ofSoundBuffer::fillWithNoise(float amplitude){
 }
 
 float ofSoundBuffer::fillWithTone( float pitchHz, float phase ){
-	float step = TWO_PI*(pitchHz/samplerate);
+	float step = glm::two_pi<float>()*(pitchHz/samplerate);
 	for (std::size_t i=0; i<size()/channels; i++ ) {
 		std::size_t base = i*channels;
 		for (std::size_t j=0; j<channels; j++)
