+++ libs/openFrameworks/types/ofBaseTypes.cpp
@@ -6,6 +6,9 @@
 #include "ofSoundBuffer.h"
 #include "ofMesh.h"
 #include "ofVideoPlayer.h"
+#include "ofSoundBaseTypes.h"
+#include "ofVideoBaseTypes.h"
+#include "ofGraphicsBaseTypes.h"
 
 using namespace std;
