+++ libs/openFrameworks/types/ofTypes.h
@@ -147,53 +147,6 @@ typedef std::unique_lock<std::mutex> ofScopedLock;
 
 
 
-/// \brief A structure describing attributes of a video format.
-///
-/// An ofVideoFormat is used to describe the size, pixel format and frame rates
-/// offered by a video device.
-///
-/// \sa ofVideoDevice
-class ofVideoFormat{
-public:
-	/// \brief The pixel format of the video format.
-	ofPixelFormat pixelFormat;
-
-	/// \brief The width of the video format in pixels.
-	int width;
-
-	/// \brief The height of the video format in pixels.
-	int height;
-
-	/// \brief A list of framerates for this video format in frames per second.
-	std::vector<float> framerates;
-};
-
-/// \brief A structure describing attributes of a video device.
-///
-/// An ofVideoDevice can represent a camera, grabber or other frame source.
-class ofVideoDevice{
-public:
-	/// \brief The video device ID.
-	int id;
-
-	/// \brief The video device name.
-	std::string deviceName;
-
-	/// \brief The video device hardware name.
-	std::string hardwareName;
-
-	/// \brief Unique identifier for the device if it has one. 
-	std::string serialID;
-
-	/// \brief A list of video device formats provided by the device.
-	/// \sa ofVideoFormat
-	std::vector<ofVideoFormat> formats;
-
-	/// \brief Is true if this video device is available.
-	bool bAvailable;
-};
-
-
 //----------------------------------------------------------
 // ofPtr
 //----------------------------------------------------------
