+++ libs/openFrameworks/video/ofGstUtils.h
@@ -3,12 +3,11 @@
 #include "ofConstants.h"
 #ifndef TARGET_ANDROID
 #include "ofConstants.h"
-#include "
-ofGraphicsBaseTypes.h"
 #include "ofPixels.h"
 #include "ofTypes.h"
 #include "ofEvents.h"
 #include "ofThread.h"
+#include "ofVideoBaseTypes.h"
 #define GST_DISABLE_DEPRECATED
 #include <gst/gst.h>
 #include <gst/gstpad.h>
